backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt==4.0.1
python-multipart>=0.0.12
psycopg2-binary>=2.9.0
alembic>=1.13.0
email-validator
apscheduler>=3.10.0
python-dotenv>=1.0.0
openai>=1.0.0
pytest>=7.0.0
httpx>=0.24.0
slowapi>=0.1.9
python-socketio>=5.0.0
